+++ tests/command/test_navigation.rs
@@ -1,4 +1,15 @@
-use branchless::testing::{make_git, GitRunOptions};
+use std::io::{Read, Write};
+use std::thread;
+use std::thread::sleep;
+use std::time::Duration;
+
+use eyre::eyre;
+
+use branchless::testing::{make_git, GitRunOptions, GitWrapper};
+use portable_pty::{native_pty_system, CommandBuilder, PtySize};
+
+const CARRIAGE_RETURN: &'static str = "\r";
+const END_OF_TEXT: &'static str = "\x03";
 
 #[test]
 fn test_prev() -> eyre::Result<()> {
@@ -192,6 +203,129 @@ fn test_next_on_master2() -> eyre::Result<()> {
         @ 70deb1e2 create test3.txt
         "###);
     }
+
+    Ok(())
+}
+
+#[test]
+#[cfg(unix)]
+fn test_checkout() -> eyre::Result<()> {
+    let git = make_git()?;
+
+    git.init_repo()?;
+    git.detach_head()?;
+    git.commit_file("test1", 1)?;
+    git.commit_file("test2", 2)?;
+    git.run(&["checkout", "master"])?;
+    git.detach_head()?;
+    git.commit_file("test3", 3)?;
+
+    run_in_pty(
+        &git,
+        &["branchless", "checkout"],
+        &["test1", CARRIAGE_RETURN],
+    )?;
+    {
+        let (stdout, _stderr) = git.run(&["smartlog"])?;
+        insta::assert_snapshot!(stdout, @r###"
+        O f777ecc9 (master) create initial.txt
+        |\
+        | @ 62fc20d2 create test1.txt
+        | |
+        | o 96d1c37a create test2.txt
+        |
+        o 98b9119d create test3.txt
+        "###);
+    }
+
+    run_in_pty(
+        &git,
+        &["branchless", "checkout"],
+        &["test3", CARRIAGE_RETURN],
+    )?;
+    {
+        let (stdout, _stderr) = git.run(&["smartlog"])?;
+        insta::assert_snapshot!(stdout, @r###"
+        O f777ecc9 (master) create initial.txt
+        |\
+        | o 62fc20d2 create test1.txt
+        | |
+        | o 96d1c37a create test2.txt
+        |
+        @ 98b9119d create test3.txt
+        "###);
+    }
+
+    Ok(())
+}
+
+#[test]
+#[cfg(unix)]
+fn test_checkout_abort() -> eyre::Result<()> {
+    let git = make_git()?;
+
+    git.init_repo()?;
+    git.detach_head()?;
+    git.commit_file("test1", 1)?;
+    git.commit_file("test2", 1)?;
+
+    run_in_pty(&git, &["branchless", "checkout"], &["test1", END_OF_TEXT])?;
+    {
+        let (stdout, _stderr) = git.run(&["smartlog"])?;
+        insta::assert_snapshot!(stdout, @r###"
+        O f777ecc9 (master) create initial.txt
+        |
+        o 62fc20d2 create test1.txt
+        |
+        @ 142901d5 create test2.txt
+        "###);
+    }
+
+    Ok(())
+}
+
+fn run_in_pty(git: &GitWrapper, args: &[&str], inputs: &[&str]) -> eyre::Result<()> {
+    // Use the native pty implementation for the system
+    let pty_system = native_pty_system();
+    let mut pty = pty_system.openpty(PtySize::default()).unwrap();
+
+    let args: Vec<&str> = {
+        let repo_path = git.repo_path.to_str().expect("Could not decode repo path");
+        let mut new_args: Vec<&str> = vec!["-C", repo_path];
+        new_args.extend(args);
+        new_args
+    };
+
+    // Spawn a git instance in the pty.
+    let mut cmd = CommandBuilder::new(&git.path_to_git);
+    let time = 0;
+    for (k, v) in git.get_base_env(&time) {
+        cmd.env(k, v);
+    }
+    cmd.args(args);
+
+    let mut child = pty
+        .slave
+        .spawn_command(cmd)
+        .map_err(|e| eyre!("Could not spawn child: {}", e))?;
+    let mut reader = pty
+        .master
+        .try_clone_reader()
+        .map_err(|e| eyre!("Could not clone reader: {}", e))?;
+
+    thread::spawn(move || loop {
+        let mut s = String::new();
+        reader.read_to_string(&mut s).unwrap();
+    });
+
+    for input in inputs {
+        // Sleep between inputs, to give the pty time to catch up.
+        sleep(Duration::from_millis(100));
+        write!(pty.master, "{}", input)?;
+        pty.master.flush()?;
+    }
+
+    child.wait()?;
 
     Ok(())
 }